api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
SQLAlchemy==2.0.27
python-dotenv==1.0.1
python-jose==3.3.0
